requirements.txt
streamlit>=1.28.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
speechrecognition>=3.10.0
pyaudio>=0.2.13
pyttsx3>=2.90
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
numpy>=1.26.2
pandas>=2.1.3
sounddevice>=0.4.6
scipy>=1.11.3